feat: Dependency Injection with Antidote 1.0.0
- Define antidote 1.0.0 validation env
- format antodite 0.7.3 with same name as antodote 1.0.0
- Note: antidote 1.0.0 is not running, DI error
