+++ antidote_1.0.0/domains/__init__.py
@@ -1,0 +1,1 @@
+"""This is the domain (entity, DTO} for exo."""
