+++ antidote_1.0.0/tests/unit/domains/test_exo.py
@@ -1,0 +1,1 @@
+"""This is the domain (entity, DTO} unit test for exo."""
