+++ antitode_0.7.3/domains/exo.py
@@ -1,0 +1,11 @@
+"""This is the domain (entity, DTO} for exo."""
+
+import dataclasses as _dtc
+
+
+@_dtc.dataclass
+class Exo:
+    """Exo DTO."""
+
+    name: str
+    description: str
