--- antitode_0.7.3/domains/nas_open_share_exception.py
@@ -1,15 +1,0 @@
-"""
-This is the DTO definition for NAS open share exception.
-
-Support:
-    Please contact hagen@devops.com
-"""
-import dataclasses as _dtc
-
-
-@_dtc.dataclass
-class NASOpenShareException:
-    """NasOpenShareException DTO."""
-
-    name: str
-    description: str
